fix (*): migrate pkg/ipfs to keeper/client/ipfs, update pkg/proof, and some data mismatches
