+++ internal/dbserver/handlers/job_read.go
@@ -16,7 +16,7 @@ func (h *Handler) GetJobsByUserAddress(c *gin.Context) {
 		return
 	}
 
-	jobIDs, err := h.userRepository.GetUserJobIDsByAddress(strings.ToLower(userAddress))
+	userID, jobIDs, err := h.userRepository.GetUserJobIDsByAddress(strings.ToLower(userAddress))
 	if err != nil {
 		h.logger.Errorf("[GetJobsByUserAddress] Error getting user data for address %s: %v", userAddress, err)
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error getting user data: " + err.Error()})
@@ -74,6 +74,6 @@ func (h *Handler) GetJobsByUserAddress(c *gin.Context) {
 
 		jobs = append(jobs, jobResponse)
 	}
-
+	h.logger.Infof("[GetJobsByUserAddress] Found %d jobs for user ID %d", len(jobs), userID)
 	c.JSON(http.StatusOK, jobs)
 }
